Initial working version with a simple server and view
This is an very early version of how a TUI could be used. It translates
the keypress events from termite into a more easy to understand format.

Currently rendering is very manual, as it requires building up a list of
instructions with escape codes, however this can be improved in future.

An example snake application is included to show how this could work for
an interactive application.