Convert Heex to a tree of boxes using LV and nimble_parsec
For ease, the LiveView Heex compiler is used to compile the templates to
strings. This gives several benefits such as syntax highlighting.
Eventually this could be vendored into the project and modified to the
needs of the project.

In line with the above, nimble_parsec is used to convert the rendered
string to a tree of boxes. Currently the styles are JSON and base64
encoded for ease, but this will be fixed in a future commit to use a
custom style format.