feat(Breeze.Server): add focusable elements
The `focusable` attribute can be set to mark an element as focusable.
This adds it to the focus tree which can be cycled using tab and
shift/tab.

To support handling actions when an element is focused, implicit state
has been added. This is stored against the `term` and passed to a
module with defined callbacks, allowing the state to be determined and
rendering to be changed depending on the implicit state.