WIP: add a scrollable viewport
An example viewport module has been added. This leverages the `offset`
style, to set the top offset. In order to calculate if an element should
be scrollable or not, the element is now passed to the implicit.

A bug that occured when the last element was an implicit has been fixed,
ensuring that the state of the last implicit is correctly tracked.